pyproject.toml
[project]
name = "neststay-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "bcrypt>=4.0.0",
    "fastapi>=0.128.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[tool.ruff]
# Target Python 3.12
target-version = "py312"

# Same as Black
line-length = 88

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "database/alembic/versions",  # Auto-generated migrations
    "tests",  # No linting on test files
]

[tool.ruff.lint]
select = [
    # Core
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "UP",     # pyupgrade (modern Python syntax)

    # Code quality
    "B",      # flake8-bugbear (common bugs)
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking (optimize imports)

    # FastAPI/Pydantic specific
    "FA",     # flake8-future-annotations
    "PIE",    # flake8-pie (misc lints)
    "RUF",    # Ruff-specific rules

    # Security
    "S",      # flake8-bandit (security)

    # Async (important for FastAPI)
    "ASYNC",  # flake8-async
]

ignore = [
    "S101",   # Allow assert (useful in tests)
    "B008",   # Allow function calls in default args (Depends() pattern)
]

[tool.ruff.lint.per-file-ignores]
# Alembic migrations are auto-generated
"database/alembic/**/*.py" = ["E501", "F401"]
# Alembic env.py needs imports after sys.path manipulation
"database/alembic/env.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["app", "database", "config"]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
# Use double quotes (consistent with Black)
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Respect magic trailing commas
skip-magic-trailing-comma = false
# Auto-detect line endings
line-ending = "auto"
# Format docstrings
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "database/alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
